airflow/scripts/git_scripts/requirements.txt
apache-airflow==2.5.0
python-dotenv==0.21.0
requests==2.28.1
pydantic==2.8.2
